hook up new preferences types to preferences system
