+++ core/src/preferences/mod.rs
@@ -1,107 +1,12 @@
 mod kv;
+mod library;
 
 pub use kv::*;
-use specta::Type;
+pub use library::*;
 
-use crate::prisma::PrismaClient;
-use crate::prisma::SortOrder;
 use std::collections::HashMap;
 
-use serde::{Deserialize, Serialize};
 use uuid::Uuid;
-
-// Preferences are a set of types that are serialized as a list of key-value pairs,
-// where nested type keys are serialized as a dot-separated path.
-// They are serailized as a list because this allows preferences to be a synchronisation boundary,
-// whereas their values (referred to as settings) will be overwritten.
-
-#[derive(Clone, Serialize, Deserialize, Type, Debug)]
-pub struct LibraryPreferences {
-	#[serde(default)]
-	#[specta(optional)]
-	location: HashMap<Uuid, LocationPreferences>,
-}
-
-impl LibraryPreferences {
-	pub async fn write(self, db: &PrismaClient) -> prisma_client_rust::Result<()> {
-		let kvs = self.to_kvs();
-
-		db._batch(kvs.into_upserts(db)).await?;
-
-		Ok(())
-	}
-
-	pub async fn read(db: &PrismaClient) -> prisma_client_rust::Result<Self> {
-		let kvs = db.preference().find_many(vec![]).exec().await?;
-
-		let prefs = PreferenceKVs::new(
-			kvs.into_iter()
-				.filter_map(|data| {
-					let a = rmpv::decode::read_value(&mut data.value?.as_slice()).unwrap();
-
-					Some((PreferenceKey::new(data.key), PreferenceValue::from_value(a)))
-				})
-				.collect(),
-		);
-
-		Ok(prefs.parse())
-	}
-}
-
-#[derive(Clone, Serialize, Deserialize, Type, Debug)]
-pub struct LocationPreferences {
-	/// View settings for the location - all writes are overwrites!
-	#[specta(optional)]
-	view: Option<LocationViewSettings>,
-}
-
-#[derive(Clone, Serialize, Deserialize, Type, Debug)]
-pub struct LocationViewSettings {
-	layout: ExplorerLayout,
-}
-
-#[derive(Clone, Serialize, Deserialize, Type, Default, Debug)]
-pub struct ListViewSettings {
-	double_click_action: Option<DoubleClickAction>,
-}
-
-pub struct MediaViewSettings {
-	item_size: Option<i32>,
-	sort_by: Option<ViewSortBy>,
-	direction: Option<SortOrder>,
-	show_square_thumbnails: Option<bool>,
-	double_click_action: Option<DoubleClickAction>,
-}
-
-pub struct GridViewSettings {
-	item_size: Option<i32>,
-	sort_by: Option<ViewSortBy>,
-	direction: Option<SortOrder>,
-	show_object_size: Option<bool>,
-	double_click_action: Option<DoubleClickAction>,
-}
-
-#[derive(Debug, Clone, Type, Serialize, Deserialize)]
-pub enum DoubleClickAction {
-	OpenFile,
-	QuickPreview,
-}
-pub enum ViewSortBy {
-	None,
-	Name,
-	Size,
-	DateCreated,
-	DateModified,
-	DateIndexed,
-	DateAccessed,
-}
-
-#[derive(Clone, Serialize, Deserialize, Type, Debug)]
-pub enum ExplorerLayout {
-	Grid,
-	List,
-	Media,
-}
 
 impl<V> Preferences for HashMap<Uuid, V>
 where
@@ -133,42 +38,10 @@ where
 	}
 }
 
-impl Preferences for LibraryPreferences {
-	fn to_kvs(self) -> PreferenceKVs {
-		let Self { location } = self;
-
-		location.to_kvs().with_prefix("location")
-	}
-
-	fn from_entries(mut entries: Entries) -> Self {
-		Self {
-			location: entries
-				.remove("location")
-				.map(|value| HashMap::from_entries(value.expect_nested()))
-				.unwrap_or_default(),
-		}
-	}
-}
-
-impl Preferences for LocationPreferences {
-	fn to_kvs(self) -> PreferenceKVs {
-		let Self { view } = self;
-
-		PreferenceKVs::new(
-			[view.map(|view| (PreferenceKey::new("view"), PreferenceValue::new(view)))]
-				.into_iter()
-				.flatten()
-				.collect(),
-		)
-	}
-
-	fn from_entries(mut entries: Entries) -> Self {
-		Self {
-			view: entries.remove("view").map(|view| view.expect_value()),
-		}
-	}
-}
-
+// Preferences are a set of types that are serialized as a list of key-value pairs,
+// where nested type keys are serialized as a dot-separated path.
+// They are serailized as a list because this allows preferences to be a synchronisation boundary,
+// whereas their values (referred to as settings) will be overwritten.
 pub trait Preferences {
 	fn to_kvs(self) -> PreferenceKVs;
 	fn from_entries(entries: Entries) -> Self;
